cleaned all code and added appropriate error handling cases and message within tests

+++ src/Part1/Test3.java
@@ -26,10 +26,17 @@ The deleted directories are 1\2, 1\2\4 and 1\2\5.  The fires deleted are:
 
 public class Test3 {
 	public static void main(String args[]){
+		if (args.length != 1)
+		{
+			System.err.println("Error. Invalid number of arguments for Test3.");
+			return;
+		}
 		Part1FS tfs = new Part1FS(TreeStorage.getTree());
 		String startingPath = args[0];
 		//maybe throw in some fail safe here?
 		tfs.deleteDirectory(startingPath);
+		System.out.println("Existing tree structure: ");
+		tfs.directory.getAllPath(tfs.directory.root);
 		TreeStorage.storeTree(tfs.directory);
 	}
 }
